# openapi_spec.yaml
openapi: 3.0.3
info:
  title: Docaiche API
  description: |
    AI-powered documentation search and content management system with advanced analytics capabilities.
    
    ## Overview
    Docaiche provides intelligent documentation search, content enrichment, and comprehensive analytics
    for tracking user behavior and system performance. The API supports real-time search with caching,
    content ingestion from multiple sources, and AI-powered content enhancement.
    
    ## Key Features
    - **Intelligent Search**: AI-powered search with technology filtering and relevance scoring
    - **Content Management**: Ingest and manage documentation from various sources
    - **Analytics Dashboard**: Comprehensive metrics for search patterns, user behavior, and system performance
    - **AI Enrichment**: Automatic content enhancement using LLM providers
    - **Real-time Monitoring**: Live system health and performance metrics
    
    ## Analytics Capabilities
    The analytics endpoints provide rich data for building beautiful dashboards:
    - Search query trends and patterns over time
    - User engagement metrics (clicks, dwell time, scroll depth)
    - Content quality scores and performance
    - System performance metrics (response times, cache hit rates)
    - Technology adoption trends
    - Real-time activity feeds
    
    ## Monitoring Integration
    - **Loki**: Centralized log aggregation for all services
...
    
  version: 2.0.0
  contact:
...
  license:
    name: MIT
...
servers:
...
paths:
...
  /search:
    get:
      tags: [search]
      summary: Search documents (GET)
      description: |
        Simple GET endpoint for search queries. Useful for browser testing and simple integrations.
      operationId: searchDocumentsGet
      parameters:
        - name: q
          in: query
          description: Search query string
          required: true
          schema:
            type: string
          example: "react hooks"
        - name: technology_hint
          in: query
          description: Filter by technology
          required: false
          schema:
            type: string
          example: "react"
        - name: limit
          in: query
          description: Maximum results to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: session_id
          in: query
          description: Session identifier for tracking
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
    
    post:
      tags: [search]
      summary: Search documents (POST)
      description: |
        Advanced search endpoint with full request body support.
        
        **Features:**
        - AI-powered relevance ranking
        - Technology-specific filtering
        - Result caching for performance
        - Background content enrichment triggers
        
        **Search Algorithm:**
        - Queries AnythingLLM vector database
        - Applies multi-factor relevance scoring
        - Considers user feedback signals
        - Optimizes for documentation-specific content
      operationId: searchDocumentsPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '500':
          description: Search execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /feedback:
    post:
      tags: [search]
      summary: Submit feedback
      description: |
        Submit user feedback on search results.
        
        **Feedback Types:**
        - helpful: Result was useful
        - not_helpful: Result wasn't useful
        - outdated: Content is outdated
        - incorrect: Content has errors
        - flag: Content needs review
        
        **Impact on System:**
        - Influences future search rankings
        - Triggers content quality reviews
        - Feeds into analytics dashboards
      operationId: submitFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '202':
          description: Feedback accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feedback received and will be processed"
  /signals:
    post:
      tags: [search]
      summary: Submit usage signal
      description: |
        Submit implicit user behavior signals.
        
        **Signal Types:**
        - click: User clicked on result
        - dwell_time: Time spent on result
        - copy: User copied content
        - share: User shared result
        - scroll_depth: How far user scrolled
        
        **Analytics Usage:**
        - Builds user engagement heatmaps
        - Tracks content effectiveness
        - Improves search relevance
      operationId: submitSignal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignalRequest'
      responses:
        '202':
          description: Signal accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Signal received and will be processed"
...
components:
  schemas:
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: Search query text
          minLength: 1
          maxLength: 500
        technology_hint:
          type: string
          description: Optional technology filter
        limit:
          type: integer
          description: Maximum results to return
          minimum: 1
          maximum: 100
          default: 20
        session_id:
          type: string
          description: Session identifier for tracking
    SearchResponse:
      type: object
      required: [results, total_count, query, execution_time_ms, cache_hit, enrichment_triggered]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total_count:
          type: integer
          description: Total matching documents
        query:
          type: string
          description: Original search query
        technology_hint:
          type: string
          description: Technology filter applied
        execution_time_ms:
          type: integer
          description: Query execution time
        cache_hit:
          type: boolean
          description: Whether results came from cache
        enrichment_triggered:
          type: boolean
          description: Whether background enrichment was triggered
    SearchResult:
      type: object
      required: [content_id, title, snippet, source_url, technology, relevance_score, content_type, workspace]
      properties:
        content_id:
          type: string
          description: Unique content identifier
        title:
          type: string
          description: Document title
        snippet:
          type: string
          description: Content preview snippet
        source_url:
          type: string
          description: Original source URL
        technology:
          type: string
          description: Technology category
        relevance_score:
          type: number
          description: Relevance score (0-1)
          minimum: 0
          maximum: 1
        content_type:
          type: string
          description: Type of content
          enum: [documentation, tutorial, api_reference, guide, example]
        workspace:
          type: string
          description: Workspace/collection name
    FeedbackRequest:
      type: object
      required: [content_id, feedback_type]
      properties:
        content_id:
          type: string
          description: Document identifier
        feedback_type:
          type: string
          enum: [helpful, not_helpful, outdated, incorrect, flag]
        rating:
          type: integer
          description: Optional rating (1-5)
          minimum: 1
          maximum: 5
        comment:
          type: string
          description: Optional comment
          maxLength: 1000
        search_query:
          type: string
          description: Associated search query
        result_position:
          type: integer
          description: Position in search results
    SignalRequest:
      type: object
      required: [content_id, signal_type]
      properties:
        content_id:
          type: string
          description: Document identifier
        signal_type:
          type: string
          enum: [click, dwell_time, copy, share, scroll_depth]
        signal_value:
          type: number
          description: Signal strength/value
        search_query:
          type: string
          description: Associated search query
        result_position:
          type: integer
          description: Position in search results
        duration_ms:
          type: integer
          description: Duration for time-based signals
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Error message
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)
security: []  # No security currently implemented